Refactor sources.py into a subpackage (#546)
