+++ pydantic_settings/__init__.py
@@ -1,3 +1,4 @@
+from .exceptions import SettingsError
 from .main import BaseSettings, CliApp, SettingsConfigDict
 from .sources import (
     CLI_SUPPRESS,
@@ -18,7 +19,6 @@ from .sources import (
     PydanticBaseSettingsSource,
     PyprojectTomlConfigSettingsSource,
     SecretsSettingsSource,
-    SettingsError,
     TomlConfigSettingsSource,
     YamlConfigSettingsSource,
     get_subcommand,
@@ -26,32 +26,32 @@ from .sources import (
 from .version import VERSION
 
 __all__ = (
-    'BaseSettings',
-    'DotEnvSettingsSource',
-    'EnvSettingsSource',
-    'CliApp',
-    'CliSettingsSource',
-    'CliSubCommand',
-    'CliSuppress',
     'CLI_SUPPRESS',
-    'CliPositionalArg',
+    'AzureKeyVaultSettingsSource',
+    'BaseSettings',
+    'CliApp',
     'CliExplicitFlag',
     'CliImplicitFlag',
     'CliMutuallyExclusiveGroup',
+    'CliPositionalArg',
+    'CliSettingsSource',
+    'CliSubCommand',
+    'CliSuppress',
+    'DotEnvSettingsSource',
+    'EnvSettingsSource',
+    'ForceDecode',
     'InitSettingsSource',
     'JsonConfigSettingsSource',
     'NoDecode',
-    'ForceDecode',
-    'PyprojectTomlConfigSettingsSource',
     'PydanticBaseSettingsSource',
+    'PyprojectTomlConfigSettingsSource',
     'SecretsSettingsSource',
     'SettingsConfigDict',
     'SettingsError',
     'TomlConfigSettingsSource',
     'YamlConfigSettingsSource',
-    'AzureKeyVaultSettingsSource',
-    'get_subcommand',
     '__version__',
+    'get_subcommand',
 )
 
 __version__ = VERSION
